+++ src/_temp_types/teamSet.ts
@@ -2,6 +2,6 @@ import {Team} from "@/_temp_types/team"
 
 export type TeamSet = {
     id: number
-    name?: string
+    name: string
     teams: Team[]
 }
