+++ src/app/students/(components)/data-table-pagination.tsx
@@ -15,9 +15,9 @@ import {
     SelectValue,
 } from "@/components/ui/select"
 
-  interface DataTablePaginationProps<TData> {
+type DataTablePaginationProps<TData> = {
     table: Table<TData>
-  }
+}
 
 export function DataTablePagination<TData>({
     table,
