+++ src/app/students/(components)/data-table-view-options.tsx
@@ -13,7 +13,7 @@ import {
     DropdownMenuSeparator,
 } from "@/components/ui/dropdown-menu"
 
-interface DataTableViewOptionsProps<TData> {
+type DataTableViewOptionsProps<TData> = {
   table: Table<TData>
 }
