+++ src/app/students/(hooks)/useDataTable.tsx
@@ -13,7 +13,7 @@ import {
 } from "@tanstack/react-table"
 
 
-interface UseDataTableProps<TData, TValue> {
+type UseDataTableProps<TData, TValue> = {
     data: TData[];
     columns: ColumnDef<TData, TValue>[];
     pageCount: number;
@@ -27,8 +27,6 @@ export function useDataTable<TData, TValue>({
     const router = useRouter()
     const searchParams = useSearchParams()
     const pathname = usePathname()
-
-    // Extract page and perPage from URL searchParams
     const page = parseInt(searchParams.get("page") as string) || 1
     const perPage = parseInt(searchParams.get("per_page") as string) || 10
 
@@ -38,7 +36,6 @@ export function useDataTable<TData, TValue>({
         pageSize: perPage,
     })
 
-    // Create query string for updating pagination
     const createQueryString = React.useCallback((params: Record<string, string | number | null>) => {
         const newSearchParams = new URLSearchParams(searchParams?.toString())
 
@@ -52,30 +49,12 @@ export function useDataTable<TData, TValue>({
 
         return newSearchParams.toString()
     }, [searchParams])
-    // // alternative createQueryString
-    // const createQueryString = (queryString:string) => {
-    //     // Use URLSearchParams for more flexible query parameter handling
-    //     const searchParams = new URLSearchParams(window.location.search)
-    //     const newParams = new URLSearchParams(queryString)
 
-    //     // Merge the new query parameters with the existing ones
-    //     newParams.forEach((value, key) => {
-    //         searchParams.set(key, value)
-    //     })
-    //     return `${searchParams.toString()}`
-    // }
-    // const updateURLAndFetchData = (queryString: string) => {
-    //     const newSearchQuery = createQueryString(queryString)
-    //     // Use the router to push the new searchQuery without reloading the page
-    //     router.push(`${pathname}?${newSearchQuery}`)
-    //     getStudentData(`http://localhost:8000/api/v1/course-members/?${newSearchQuery}`)
-    // }
     const pagination = React.useMemo(() => ({
         pageIndex,
         pageSize,
     }),
     [pageIndex, pageSize])
-    // Update the URL based on pagination changes
     React.useEffect(() => {
         router.push(`${pathname}?${createQueryString({
             page: pageIndex + 1,
@@ -86,7 +65,6 @@ export function useDataTable<TData, TValue>({
         })
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [pageIndex, pageSize])
-    // Initialize React Table with server-side pagination
     const table = useReactTable({
         data,
         columns,
