+++ src/app/students/page.tsx
@@ -1,12 +1,12 @@
 'use client'
 import React from 'react'
-import { StudentsProvider } from './hooks/useStudents'
-import StudentsPage from '@/app/students/views/studentsPage'
+import { StudentsProvider } from './(hooks)/useStudents'
+import { StudentsTable } from './(components)/students-table'
 
-export default function page() {
+export default function StudentsPage() {
     return (
         <StudentsProvider>
-            <StudentsPage />
+            <StudentsTable/>
         </StudentsProvider>
     )
 }
